refactor: Move SalesChannelProductEntity logic into SalesChannelProductSubscriber
