+++ src/Core/Content/DependencyInjection/product.xml
@@ -5,12 +5,6 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="Shopware\Core\Content\Product\SalesChannel\SalesChannelProductSubscriber">
-            <tag name="kernel.event_subscriber"/>
-            <argument type="service" id="Shopware\Core\System\SystemConfig\SystemConfigService"/>
-            <argument type="service" id="Shopware\Core\Content\Product\SalesChannel\Price\ProductPriceCalculator"/>
-        </service>
-
         <service id="Shopware\Core\Content\Product\DataAbstractionLayer\ProductExceptionHandler">
             <tag name="shopware.dal.exception_handler"/>
         </service>
@@ -142,7 +136,16 @@
             <argument type="service" id="event_dispatcher"/>
         </service>
 
+        <service id="Shopware\Core\Content\Product\SalesChannelProductBuilder">
+            <argument type="service" id="Shopware\Core\System\SystemConfig\SystemConfigService"/>
+        </service>
+
+        <service id="Shopware\Core\Content\Product\ProductVariationBuilder"/>
+
         <service id="Shopware\Core\Content\Product\Subscriber\ProductSubscriber">
+            <argument type="service" id="Shopware\Core\Content\Product\SalesChannelProductBuilder"/>
+            <argument type="service" id="Shopware\Core\Content\Product\ProductVariationBuilder"/>
+            <argument type="service" id="Shopware\Core\Content\Product\SalesChannel\Price\ProductPriceCalculator"/>
             <tag name="kernel.event_subscriber"/>
         </service>
