+++ src/Core/Content/Product/Subscriber/ProductSubscriber.php
@@ -5,23 +5,38 @@ namespace Shopware\Core\Content\Product\Subscriber;
 use Shopware\Core\Content\Product\DataAbstractionLayer\CheapestPrice\CheapestPriceContainer;
 use Shopware\Core\Content\Product\ProductEntity;
 use Shopware\Core\Content\Product\ProductEvents;
+use Shopware\Core\Content\Product\ProductVariationBuilder;
+use Shopware\Core\Content\Product\SalesChannel\Price\AbstractProductPriceCalculator;
 use Shopware\Core\Content\Product\SalesChannel\SalesChannelProductEntity;
-use Shopware\Core\Content\Property\Aggregate\PropertyGroupOption\PropertyGroupOptionCollection;
-use Shopware\Core\Content\Property\Aggregate\PropertyGroupOption\PropertyGroupOptionEntity;
-use Shopware\Core\Content\Property\PropertyGroupCollection;
+use Shopware\Core\Content\Product\SalesChannelProductBuilder;
 use Shopware\Core\Framework\DataAbstractionLayer\Event\EntityLoadedEvent;
 use Shopware\Core\Framework\Feature;
+use Shopware\Core\System\SalesChannel\Entity\SalesChannelEntityLoadedEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class ProductSubscriber implements EventSubscriberInterface
 {
+    private SalesChannelProductBuilder $salesChannelProductBuilder;
+
+    private ProductVariationBuilder $productVariationBuilder;
+
+    private AbstractProductPriceCalculator $calculator;
+
+    public function __construct(
+        SalesChannelProductBuilder $salesChannelProductBuilder,
+        ProductVariationBuilder $productVariationBuilder,
+        AbstractProductPriceCalculator $calculator
+    ) {
+        $this->salesChannelProductBuilder = $salesChannelProductBuilder;
+        $this->productVariationBuilder = $productVariationBuilder;
+        $this->calculator = $calculator;
+    }
+
     public static function getSubscribedEvents(): array
     {
-        // Return the events to listen to as array like this:  <event to listen to> => <method to execute>
         return [
-            ProductEvents::PRODUCT_LOADED_EVENT => [
-                ['loaded'],
-            ],
+            ProductEvents::PRODUCT_LOADED_EVENT => 'loaded',
+            'sales_channel.' . ProductEvents::PRODUCT_LOADED_EVENT => 'salesChannelLoaded',
         ];
     }
 
@@ -40,93 +55,32 @@ class ProductSubscriber implements EventSubscriberInterface
                 }
             }
 
-            $variants = $product->getOptions() === null ? [] : $this->buildVariation($product->getOptions()->getElements());
-            $product->setVariation($variants);
+            $this->productVariationBuilder->build($product);
+        }
+    }
 
-            if ($product instanceof SalesChannelProductEntity) {
-                if (Feature::isActive('FEATURE_NEXT_16151')) {
-                    $price = $product->getCheapestPrice();
+    public function salesChannelLoaded(SalesChannelEntityLoadedEvent $event): void
+    {
+        $context = $event->getSalesChannelContext();
 
-                    if ($price instanceof CheapestPriceContainer) {
-                        $resolved = $price->resolve($event->getContext());
-                        $product->setCheapestPriceContainer($price);
-                        $product->setCheapestPrice($resolved);
-                    }
+        /** @var SalesChannelProductEntity $product */
+        foreach ($event->getEntities() as $product) {
+            if (Feature::isActive('FEATURE_NEXT_16151')) {
+                $price = $product->getCheapestPrice();
+
+                if ($price instanceof CheapestPriceContainer) {
+                    $resolved = $price->resolve($event->getContext());
+                    $product->setCheapestPriceContainer($price);
+                    $product->setCheapestPrice($resolved);
                 }
-                $product->setSortedProperties(
-                    $this->sortProperties($product)
-                );
             }
-        }
-    }
 
-    private function sortProperties(SalesChannelProductEntity $product): PropertyGroupCollection
-    {
-        $properties = $product->getProperties();
-        if ($properties === null) {
-            return new PropertyGroupCollection();
+            $this->salesChannelProductBuilder->build($product, $context);
         }
 
-        $sorted = [];
-        foreach ($properties as $option) {
-            $origin = $option->getGroup();
-
-            if (!$origin || !$origin->getVisibleOnProductDetailPage()) {
-                continue;
-            }
-            $group = clone $origin;
-
-            $groupId = $group->getId();
-            if (\array_key_exists($groupId, $sorted)) {
-                \assert($sorted[$groupId]->getOptions() !== null);
-                $sorted[$groupId]->getOptions()->add($option);
-
-                continue;
-            }
-
-            if ($group->getOptions() === null) {
-                $group->setOptions(new PropertyGroupOptionCollection());
-            }
-
-            \assert($group->getOptions() !== null);
-            $group->getOptions()->add($option);
-
-            $sorted[$groupId] = $group;
-        }
-
-        $collection = new PropertyGroupCollection($sorted);
-        $collection->sortByPositions();
-        $collection->sortByConfig();
-
-        return $collection;
-    }
-
-    private function buildVariation(array $options): array
-    {
-        uasort($options, function (PropertyGroupOptionEntity $a, PropertyGroupOptionEntity $b) {
-            if ($a->getGroup() === null || $b->getGroup() === null) {
-                return $a->getGroupId() <=> $b->getGroupId();
-            }
-
-            if ($a->getGroup()->getPosition() === $b->getGroup()->getPosition()) {
-                return $a->getGroup()->getTranslation('name') <=> $b->getGroup()->getTranslation('name');
-            }
-
-            return $a->getGroup()->getPosition() <=> $b->getGroup()->getPosition();
-        });
-
-        // fallback - simply take all option names unordered
-        $names = array_map(function (PropertyGroupOptionEntity $option) {
-            if (!$option->getGroup()) {
-                return [];
-            }
-
-            return [
-                'group' => $option->getGroup()->getTranslation('name'),
-                'option' => $option->getTranslation('name'),
-            ];
-        }, $options);
-
-        return array_values($names);
+        $this->calculator->calculate(
+            $event->getEntities(),
+            $event->getSalesChannelContext()
+        );
     }
 }
